Refactor to have common code in cdk and python for both rest and ws apis. Get api keys and usage plans working for ws api
