+++ cdk/template.env
@@ -3,7 +3,6 @@ API_GATEWAY_USE_API_KEY="true"   # "true" or "false"
 API_GATEWAY_USE_IAM_AUTH="true"  # "true" or "false"
 DEFAULT_MAX_TOKENS="4096"
 DEFAULT_TEMP="0.0"
-ECR_WEBSOCKET_REPOSITORY="llm-gateway-ws"
 UI_CERT_ARN=""
 UI_DOMAIN_NAME=""
 ECR_STREAMLIT_REPOSITORY="mystreamlitrepo"
@@ -19,3 +18,4 @@ AZURE_OPENAI_ENDPOINT=""
 AZURE_OPENAI_API_KEY=""
 AZURE_OPENAI_API_VERSION="2024-02-01"
 ECR_API_KEY_REPOSITORY=""
+ECR_LLM_GATEWAY_REPOSITORY="llm-gateway-repo"
