+++ lambdas/gateway/.gitignore
