+++ lambdas/gateway/app_rest.py
