+++ lambdas/gateway/requirements.txt
